Make optional to check for missing movescount entries ID on startup
Just after laaunch the app, it'll check for missing movescount activity
IDs, by checking all the local hosts, and downloading the needed
activities from movescount.

This operation takes a long time, making high usage of the CPU.

Added a checkbox to allow the user choose if sync it or not.